backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.10,<4.0
djongo>=1.3.6